nptl: sh: Fix Race conditions in pthread cancellation (BZ#12683)
This patch adds the sh modifications required for the BZ#12683.
It basically adds the required __pthread_get_pc function.

SH requires an arch-specific syscall_cancel because the
INTERNAL_SYSCALL_NCS adds the required or instruction to workaround
a hardware bug.  The implementation were based on on default C
version built with GCC 6.1

Checked against a build and make check run-built-tests=no for
sh4-linux-gnu.

	* sysdeps/unix/sysv/linux/sh/sysdep-cancel.h (__pthread_get_pc):
	New function.
	* sysdeps/unix/sysv/linux/sh/syscall_cancel.S: New file.